Introduce Fixtures to index test resources correctly
