+++ plugin/functional-tests/src/test/groovy/com/novoda/gradle/test/BuildFolder.groovy
@@ -2,37 +2,14 @@ package com.novoda.gradle.test
 
 class BuildFolder {
 
-    private Closure<File> rootDirProvider
+    private final File rootDir
 
     BuildFolder(String path = '') {
-        def start = new File(getResource('.').file)
-        if (start.path.endsWith('build/classes/groovy/test')) {
-            rootDir(new File(start.parentFile.parentFile.parentFile, path))
-        } else if (start.path.endsWith('out/test/classes')) {
-            rootDir(new File(start.parentFile.parentFile.parentFile, "build/$path"))
-        } else {
-            throw new UnsupportedOperationException("Unable to identify build folder from path: $start")
-        }
-    }
-
-    private void rootDir(File file) {
-        rootDirProvider = {
-            file.mkdirs()
-            return file
-        }.memoize()
-    }
-
-    private static URL getResource(String resourceName) {
-        ClassLoader loader = Thread.currentThread().getContextClassLoader() ?: BuildFolder.class.getClassLoader()
-        URL url = loader.getResource(resourceName)
-        if (url == null) {
-            throw new IllegalArgumentException("resource ${resourceName} not found.")
-        }
-        return url
+        this.rootDir = new File(Fixtures.BUILD_DIR, path)
     }
 
     File getRootDir() {
-        rootDirProvider.call()
+        rootDir
     }
 
     File newFolder(String path) {
